Move validate property to MetadataOptions
