+++ components/formats-api/src/loci/formats/FormatHandler.java
@@ -32,8 +32,15 @@
 
 package loci.formats;
 
+import java.util.HashSet;
+import java.util.Set;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import loci.formats.in.DefaultMetadataOptions;
+import loci.formats.in.MetadataLevel;
+import loci.formats.in.MetadataOptions;
 
 /**
  * Abstract superclass of all biological file format readers and writers.
@@ -59,8 +66,8 @@ public abstract class FormatHandler implements IFormatHandler {
   /** Name of current file. */
   protected String currentId;
   
-  /** Whether or not to validate files. */
-  protected boolean validate = true;
+  /** Metadata parsing options. */
+  protected MetadataOptions metadataOptions = new DefaultMetadataOptions();
 
   // -- Constructors --
 
@@ -73,6 +80,36 @@ public abstract class FormatHandler implements IFormatHandler {
   public FormatHandler(String format, String[] suffixes) {
     this.format = format;
     this.suffixes = suffixes == null ? new String[0] : suffixes;
+  }
+  
+  // -- IMetadataConfigurable API methods --
+
+  /* (non-Javadoc)
+   * @see loci.formats.IMetadataConfigurable#getSupportedMetadataLevels()
+   */
+  @Override
+  public Set<MetadataLevel> getSupportedMetadataLevels() {
+    Set<MetadataLevel> supportedLevels = new HashSet<MetadataLevel>();
+    supportedLevels.add(MetadataLevel.ALL);
+    supportedLevels.add(MetadataLevel.NO_OVERLAYS);
+    supportedLevels.add(MetadataLevel.MINIMUM);
+    return supportedLevels;
+  }
+
+  /* (non-Javadoc)
+   * @see loci.formats.IMetadataConfigurable#getMetadataOptions()
+   */
+  @Override
+  public MetadataOptions getMetadataOptions() {
+    return metadataOptions;
+  }
+
+  /* (non-Javadoc)
+   * @see loci.formats.IMetadataConfigurable#setMetadataOptions(loci.formats.in.MetadataOptions)
+   */
+  @Override
+  public void setMetadataOptions(MetadataOptions options) {
+    this.metadataOptions = options;
   }
 
   // -- IFormatHandler API methods --
@@ -100,18 +137,6 @@ public abstract class FormatHandler implements IFormatHandler {
   public Class<?> getNativeDataType() {
     // NB: Handlers use byte arrays by default as the native type.
     return byte[].class;
-  }
-  
-  /* @see IFormatHandler#setValidate(boolean) */
-  @Override
-  public void setValidate(boolean validateFile) {
-    validate = validateFile;
-  }
-
-  /* @see IFormatHandler#isValidate() */
-  @Override
-  public boolean isValidate() {
-    return validate;
   }
 
   // -- Utility methods --
