+++ components/formats-api/src/loci/formats/IFormatHandler.java
@@ -38,7 +38,7 @@ import java.io.IOException;
 /**
  * Interface for all biological file format readers and writers.
  */
-public interface IFormatHandler extends Closeable {
+public interface IFormatHandler extends Closeable, IMetadataConfigurable {
 
   /** Checks if the given string is a valid filename for this file format. */
   boolean isThisType(String name);
@@ -60,11 +60,5 @@ public interface IFormatHandler extends Closeable {
 
   /** Sets the current file name. */
   void setId(String id) throws FormatException, IOException;
-  
-  /** Specifies whether or not to validate files when reading. */
-  void setValidate(boolean validate);
-
-  /** Returns true if files should be validated when read.*/
-  boolean isValidate();
 
 }
