+++ components/formats-api/src/loci/formats/IFormatReader.java
@@ -42,7 +42,7 @@ import loci.formats.meta.MetadataStore;
 /**
  * Interface for all biological file format readers.
  */
-public interface IFormatReader extends IFormatHandler, IMetadataConfigurable {
+public interface IFormatReader extends IFormatHandler {
 
   // -- Constants --
